# CassandraService::getKnownPort uses intStream