--- a/atlasdb-cassandra/src/main/java/com/palantir/atlasdb/keyvalue/cassandra/pool/CassandraService.java
+++ b/atlasdb-cassandra/src/main/java/com/palantir/atlasdb/keyvalue/cassandra/pool/CassandraService.java
@@ -17,7 +17,6 @@ package com.palantir.atlasdb.keyvalue.cassandra.pool;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Suppliers;
-import com.google.common.collect.FluentIterable;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableRangeMap;
@@ -58,13 +57,13 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.PrimitiveIterator.OfInt;
 import java.util.Random;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
@@ -73,6 +72,7 @@ import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 import java.util.stream.Stream;
 import org.apache.cassandra.thrift.EndpointDetails;
 import org.apache.cassandra.thrift.TokenRange;
@@ -302,22 +302,23 @@ public class CassandraService implements AutoCloseable {
     }
 
     private int getKnownPort() throws UnknownHostException {
-        // explicitly not using streams due to allocation overhead
-        return onlyPort(FluentIterable.concat(
-                        FluentIterable.from(currentPools.keySet()).transform(CassandraServer::proxy),
-                        getServersSocketAddressesFromConfig())
-                .transform(InetSocketAddress::getPort));
+        // explicitly not collecting to a set to avoid allocation overhead
+        return onlyPort(Stream.concat(
+                        currentPools.keySet().stream().map(CassandraServer::proxy),
+                        getServersSocketAddressesFromConfig().stream())
+                .mapToInt(InetSocketAddress::getPort));
     }
 
     @VisibleForTesting
-    static int onlyPort(Iterable<? extends Integer> iterable) throws UnknownHostException {
-        Iterator<? extends Integer> iterator = iterable.iterator();
+    static int onlyPort(IntStream intStream) throws UnknownHostException {
+        // explicitly avoiding boxing from int -> Integer
+        OfInt iterator = intStream.iterator();
         if (!iterator.hasNext()) {
             throw new UnknownHostException("No single known port");
         }
-        int port = iterator.next();
+        int port = iterator.nextInt();
         while (iterator.hasNext()) {
-            if (port != iterator.next()) {
+            if (port != iterator.nextInt()) {
                 throw new UnknownHostException("No single known port");
             }
         }