## General
**Before this PR**:
https://github.com/palantir/atlasdb/pull/6451 attempted to reduce allocations and overhead of `CassandraService::getKnownPort`; however, some JFR are showing significant allocations from boxed `int` -> `Integer`.

**After this PR**:
<!-- User-facing outcomes this PR delivers go below -->
==COMMIT_MSG==
CassandraService::getKnownPort uses intStream

```
Benchmark                                    (size)  Mode  Cnt     Score      Error   Units
PortBenchmakr.boxedStream                         3  avgt    3    89.652 ±   23.727   ns/op
PortBenchmakr.boxedStream:·gc.alloc.rate.norm     3  avgt    3   464.031 ±    0.020    B/op
PortBenchmakr.intStream                           3  avgt    3    61.106 ±    0.765   ns/op
PortBenchmakr.intStream:·gc.alloc.rate.norm       3  avgt    3   416.023 ±    0.017    B/op
PortBenchmakr.iterator                            3  avgt    3    12.173 ±    0.578   ns/op
PortBenchmakr.iterator:·gc.alloc.rate.norm        3  avgt    3    96.005 ±    0.003    B/op

PortBenchmakr.boxedStream                        30  avgt    3   505.253 ±  129.755   ns/op
PortBenchmakr.boxedStream:·gc.alloc.rate.norm    30  avgt    3   896.081 ±    0.302    B/op
PortBenchmakr.intStream                          30  avgt    3   356.273 ±   24.996   ns/op
PortBenchmakr.intStream:·gc.alloc.rate.norm      30  avgt    3   416.039 ±    0.200    B/op
PortBenchmakr.iterator                           30  avgt    3    29.301 ±    3.872   ns/op
PortBenchmakr.iterator:·gc.alloc.rate.norm       30  avgt    3    96.008 ±    0.008    B/op
```

==COMMIT_MSG==

**Priority**:

**Concerns / possible downsides (what feedback would you like?)**:
In microbenchmarks above, `IntStream` somewhat surprisingly has worse time and allocation overhead than a boxed `Iterator<Integer>`.

**Is documentation needed?**:

## Compatibility
**Does this PR create any API breaks (e.g. at the Java or HTTP layers) - if so, do we have compatibility?**:

**Does this PR change the persisted format of any data - if so, do we have forward and backward compatibility?**:

**The code in this PR may be part of a blue-green deploy. Can upgrades from previous versions safely coexist? (Consider restarts of blue or green nodes.)**:

**Does this PR rely on statements being true about other products at a deployment - if so, do we have correct product dependencies on these products (or other ways of verifying that these statements are true)?**:

**Does this PR need a schema migration?**

## Testing and Correctness
**What, if any, assumptions are made about the current state of the world? If they change over time, how will we find out?**:

**What was existing testing like? What have you done to improve it?**:

**If this PR contains complex concurrent or asynchronous code, is it correct? The onus is on the PR writer to demonstrate this.**:

**If this PR involves acquiring locks or other shared resources, how do we ensure that these are always released?**:

## Execution
**How would I tell this PR works in production? (Metrics, logs, etc.)**:

**Has the safety of all log arguments been decided correctly?**:

**Will this change significantly affect our spending on metrics or logs?**:

**How would I tell that this PR does not work in production? (monitors, etc.)**:

**If this PR does not work as expected, how do I fix that state? Would rollback be straightforward?**:

**If the above plan is more complex than “recall and rollback”, please tag the support PoC here (if it is the end of the week, tag both the current and next PoC)**:

## Scale
**Would this PR be expected to pose a risk at scale? Think of the shopping product at our largest stack.**:

**Would this PR be expected to perform a large number of database calls, and/or expensive database calls (e.g., row range scans, concurrent CAS)?**:

**Would this PR ever, with time and scale, become the wrong thing to do - and if so, how would we know that we need to do something differently?**:

## Development Process
**Where should we start reviewing?**:

**If this PR is in excess of 500 lines excluding versions lock-files, why does it not make sense to split it?**:

**Please tag any other people who should be aware of this PR**:
@jeremyk-91
@sverma30
@raiju

<!---
Please remember to:
- Add any necessary release notes (including breaking changes)
- Make sure the documentation is up to date for your change
--->