@schlosna 2023-03-10
closing as `IntStream` (surprisingly) does not seem more efficient